fix: change file name and move tests into test folder
