+++ src/Constants.ts
@@ -1,4 +1,4 @@
-import { ITask } from './Interfaces';
+import { ITask } from './interfaces';
 export const tasks: ITask[] = [
   { id: 1, name: 'task 01', completed: false },
   { id: 2, name: 'task 02', completed: true },
