+++ src/components/__tests__/TodoForm.test.tsx
@@ -1,5 +1,5 @@
 import { fireEvent, render, screen } from '@testing-library/react';
-import TodoForm from './TodoForm';
+import TodoForm from '../TodoForm';
 import '@testing-library/jest-dom';
 import userEvent from '@testing-library/user-event';
